requirements.txt
django-filter==24.2
djangorestframework==3.15.1
Markdown==3.6
